elf32: Stack top should start 8 bytes ahead to prevent buffer overflows
This issue was triggering all kind of weird random errors when restoring
the context of a finishing process. The stack was overflowing the canary
value and extending one extra byte on some non-allocated heap memory.
Therefore, any attempt to use malloc() might screw the return address saved
in the stack